Analytics/requirements.txt
mysql-connector-python
pandas
scikit-learn
numpy